requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
keras
flask
requests
mysql-connector-python
pymysql
python-dotenv
catboost
xgboost
tqdm
# -e . # This line is used for editable installations(you can use this if you just want to use "pip install -r requirements.txt" instead of "python setup.py install") and should not be included in the requirements.txt file.